+++ core/src/main/java/org/sterl/spring/persistent_tasks/trigger/event/TriggerAddedEvent.java
@@ -1,0 +1,7 @@
+package org.sterl.spring.persistent_tasks.trigger.event;
+
+import org.sterl.spring.persistent_tasks.trigger.model.TriggerEntity;
+
+public record TriggerAddedEvent(TriggerEntity trigger) implements TriggerLifeCycleEvent {
+
+}