+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
 ## [v1.0.1]
+
+- Showing trigger history entries
+- Added PersistentTaskService as a abstraction